Add support for K64F
This commit adds support for building the MCUboot demo for the K64F target. Note that you will need to insert an SD card into the on-board slot for the demo to work.

Also note that if you are using the shared data feature, you will need to use the `-l` flag when executing `mbed compile` to provide the customized linker script in linker/TARGET_K64F/TOOLCHAIN_GCC_ARM. This linker script reserves a region of RAM at the end of physical RAM of the size set by `mcuboot.share-data-size`.